% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_test.R
\name{bayesian_test}
\alias{bayesian_test}
\title{Highlights and test the time-course of effects through bayes factors}
\usage{
bayesian_test(data, formula, time, whichRandom, whichModels = "bottom", ...)
}
\arguments{
\item{data}{A data.frame containing all the necessary variables.}

\item{formula}{A 'lme4'-style formula, passed as a string.}

\item{time}{A string indicating the name of the time variable.}

\item{whichRandom}{A string (vector) indicating which variables in the formula are
random effects. Random effects must be factors. see '?anovaBF'}

\item{whichModels}{defaults to "bottom", meaning a type 2 test
of inclusion. see '?anovaBF'}

\item{...}{Further arguments that are passed to 'BayesFactor::generalTestBF()'.
For example in cases in which you intend to deviate from default priors.}
}
\value{
A list including, for each term in the model, the bayes factors for
each timepoint.
}
\description{
This function performs general  bayesian tests through
'BayesFactor::generalTestBF()'.
}
