% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_series.R
\name{check_all_series}
\alias{check_all_series}
\title{Convenience function to check series across IDs and Trials, and save a plot in the current path}
\usage{
check_all_series(data, ID, Trial, series1, series2, time)
}
\arguments{
\item{data}{Mandatory, differently from 'check_series'. The IDs and Trials
levels around which the loop is set are retrieved here.}

\item{ID}{A string indicating the name of the ID column.}

\item{Trial}{A string indicating the name of the Trial column.}

\item{series1}{Unlike 'check_series', this must only be a string
indicating the name of the first time series to plot.}

\item{series2}{Unlike 'check_series', this must only be a string
indicating the name of the second time series to plot.}

\item{time}{Unlike 'check_series', this must only be a string
indicating the elapsed time, which will be used for the x-axis.}
}
\value{
A plot.
}
\description{
Convenience function to check series across IDs and Trials, which saves multiple
plots in the current path by creating folders and subfolders in a hopefully
meaningful way.
Note that with massive data this could take some time, so you may want to
debug first with a subset of the data.
The defaults for 'ggsave', which is used internally, are for now a bit stiff but may become more
flexible in the future.
The 'check_series' function is invoked for plots, thus you may want to check
the relative help page.
}
