% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_series.R
\name{check_series}
\alias{check_series}
\title{Plots two time series against each other}
\usage{
check_series(data, series1, series2, time)
}
\arguments{
\item{data}{Optional. Can be omitted if passed through dplyr's style
pipelines (through '.'), in which case the other arguments should be
passed as quoted variables' names.}

\item{series1}{A vector variable with values for the first time series.
It is plotted by means of a black line. Typically, the original data.}

\item{series2}{A vector variable with values for the second time series.
It is plotted by means of a red line. Typically, the processed data.}

\item{time}{A vector variable indicating the elapsed time, which is
used for the x axis.}
}
\value{
A plot.
}
\description{
The function plots two time series against each other.
The intended use is generally to check the original vs. preprocessed data.
The first series is plotted with a black line, the second one - typically
the reconstructed series - with a red line. Nas show up as interruptions
of the lines. The function can be used within dplyr's style pipes -
in which case 'data' can be omitted and the other variables must
be provided as quoted variables' names - or standard vectors may be
provided - in which case 'data' should be NULL and args should be passed
by name.
}
