% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_signal.R
\name{consolidate_signal}
\alias{consolidate_signal}
\title{Consolidate pupil data according to different heuristics}
\usage{
consolidate_signal(
  s1,
  s2,
  v1,
  v2,
  strategy = c("conservative", "liberal", "pick_best"),
  plausible = NULL
)
}
\arguments{
\item{s1}{A vector with the first signal.}

\item{s2}{A vector with the second signal.}

\item{v1}{A vector with the weights for the first signal.}

\item{v2}{A vector with the weights for the second signal.}

\item{strategy}{A strategy for mixing the two signals. Conservative takes
the mean only when both signals are valid. Liberal additionally take
one valid signal even when the other is not valid. Pick_best chooses the
best overall signal (valid more often) and disregard the other when
only one of the two is valid.}

\item{plausible}{A vector of length 2 defining the range of plausible values
for pupil size. If provided, values outside this range are set to NA.}
}
\value{
A numeric vector with the consolidated pupil size or NAs when not
available.
}
\description{
Mostly used, e.g., when pupil data is available for both eyes and one needs
a single variable. Results are weighted by vectors of signal validity
as provided by the eye-trackers. If absent, that all signals is valid is assumed.
}
