% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_signal_g.R
\name{decode_signal_g}
\alias{decode_signal_g}
\title{Highlights and test the time-course of effects through crossvalidation}
\usage{
decode_signal_g(
  data,
  formula,
  dv,
  time,
  id,
  trial,
  nfolds = 4,
  t_thresh = 2,
  consensus_thresh = 0.75,
  ...
)
}
\arguments{
\item{data}{A data.frame containing all the necessary variables.}

\item{formula}{A 'lme4'-style formula, passed as a string.}

\item{dv}{A string indicating the name of the dependent variable.}

\item{time}{A string indicating the name of the time variable.}

\item{id}{A string indicating the name of the id (participant) variable.}

\item{trial}{A string indicating the name of the trial variable.}

\item{nfolds}{Number of folds to split trials in. Defaults to 4.}

\item{t_thresh}{Used to seek consensus: the minimum t-value required to push the
time-point forward.}

\item{consensus_thresh}{The minimum proportion of time-points that must be above
't_thresh' across folds in order to keep the time-point in the consensus.}

\item{...}{Further params for 'glmer()', e.g. "family".}
}
\value{
A list including: peaks retained for each (left-out) fold; test of the
retained, cross-validated peaks; test of the consensus time-points, if any; list
of time-points retained in the consensus for each effect.
}
\description{
Same as 'decode_signal()' except it is powered by
'glmer()' and thus performs generalized LMEMs.
}
