% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_bayes.R
\name{interpret_bayes}
\alias{interpret_bayes}
\title{Interprets easily the results of 'bayesian_test'}
\usage{
interpret_bayes(bfs, evidence_thresh = 10, cluster_size = 10)
}
\arguments{
\item{bfs}{An object returned by 'bayesian_test'.}

\item{evidence_thresh}{Threshold to judge evidence. Defaults
to 10. It is meant to be symmetrical: values above 10
indicate support for the alternative Hp, less than 1/10 for the null.}

\item{cluster_size}{How many timepoints constitute a cluster?}
}
\value{
A list including intervals that support H0 or H1.
}
\description{
This is a helper function. It takes an object
created by 'bayesian_test' and spits intervals
with evidence for either the null or alternative models.
}
