% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_lmem.R
\name{interpret_lmem}
\alias{interpret_lmem}
\title{Interprets easily the results of 'lmem_test'}
\usage{
interpret_lmem(lmems, p = "p.corrected", sig_thresh = 0.05, cluster_size = 10)
}
\arguments{
\item{lmems}{An object returned by 'lmem_test'.}

\item{p}{defaults to "corrected" to use adjusted p values}

\item{sig_thresh}{Threshold to judge significance.}

\item{cluster_size}{How many timepoints constitute a cluster?}
}
\value{
A list including intervals with significant fixed effects.
}
\description{
This is a helper function. It takes an object
created by 'lmem_test' and spits intervals
with significant effects.
}
