% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lmem.R
\name{plot_lmem}
\alias{plot_lmem}
\title{takes an LMEM object and plots the results}
\usage{
plot_lmem(LMEM, exclude_intercept = T, subset_significant = T)
}
\arguments{
\item{LMEM}{A lmem_test object.}

\item{exclude_intercept}{Whether the intercept should be excluded.}

\item{subset_significant}{Whether only significant clusters should be plotted.}
}
\value{
A plot powered by 'ggplot2'.
}
\description{
This function is meant to depict the time course of
t statistics for each effect in a lmem_test object. At state it is
very essential, and there is little room to set graphical pars from
within this function.
}
