% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot loadings of a reduced time-series}
\usage{
plot_loadings(name, data)
}
\arguments{
\item{name}{A string indicating the component to depict (e.g., "PC1").}

\item{data}{An object as returned by, e.g., 'reduce_PCA'.}
}
\value{
A plot powered by 'ggplot2'.
}
\description{
This function is under active development. It is
meant to depict the loadings of components obtained by reducing
pupillary time-series, as to ease interpretation.
}
