% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manifold.R
\name{plot_manifold}
\alias{plot_manifold}
\title{Plot a 3D scatterplot (based on plot3D) based on a reduce_rPCA object}
\usage{
plot_manifold(
  Scores,
  colvar = NULL,
  col = NULL,
  theta = 0,
  phi = 0,
  pch = 16,
  cex = 2,
  adapt3D = F,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL
)
}
\arguments{
\item{Scores}{An object as returned by 'reduce_rPCA'.}

\item{adapt3D}{Whether axis should be "as is", FALSE, or strectched to fit
the 3D box, TRUE}

\item{colvar;col}{The name of a variable to use for colors;
a char vectors of colors.}

\item{theta;phi}{The angles of the "eyes" on the plot}

\item{pch;cex}{Some graphical pars defining shape and size
of the points plotted}

\item{xlim;ylim;zlim}{To manually supply limits}
}
\value{
A 3D plot powered by 'plot3D'.
}
\description{
This function is under active development. It is
meant to depict the three scores of components obtained by reducing
pupillary time-series through rPCA.
There are some strong assumptions, first and foremost that the object is
from rotated PCA, and that it includes variables named
"RC1" - x axis, "RC2" - y axis, and "RC3" - z axis.
The function is a wrapper around 'plot3D::scatter3D'
and thus the package is required.
}
