% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_options.R
\name{pp_options}
\alias{pp_options}
\title{Set or get options for Pupilla's preprocessing parameters}
\usage{
pp_options(...)
}
\arguments{
\item{thresh}{\strong{speed_clean} Threshold (z point or absolute value) above which values are marked as NA.}

\item{speed_method}{\strong{speed_clean} Whether the 'thresh' is a z-score ('z'), with deviant values
omitted once, or until there are no more values above the threshold ('z-dynamic').
'abs' is used instead when a precise absolute value for speed is supplied.}

\item{extend_by}{\strong{speed_clean} Number of samples starting from the deviant speed values that
are stripped (e.g., the signal in proximity of blinks may be biased as well).}

\item{island_size}{\strong{speed_clean} Islands of signal in the midst of NAs are removed if smaller
or equal to this threshold (amount of samples).}

\item{extend_blink}{\strong{interpolate} NAs are extended by these many samples
prior to interpolation. This gets rid of signal that may be compromised
in the close proximity of a blink.}

\item{overall_thresh}{\strong{interpolate} Overall quality threshold: e.g., the total amount of data
that is allowed to be missing from the original vector.}

\item{consecutive_thresh}{\strong{interpolate} Consecutive gaps in the signal: e.g., the total amount of data
that is allowed to be missing from the original vector \strong{consecutively}.}

\item{disable_smoothing}{\strong{smooth} Whether you want to disable smoothing. This parameter
is FALSE by default; if you want to disable the smoothing, set it to TRUE.}

\item{spar}{\strong{smooth} Smoothing factor as in 'smooth.spline()'.}
}
\description{
Set or get options for Pupilla's preprocessing parameters
}
