% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\title{A convenience function to preprocess pupillometry data}
\usage{
pre_process(vector, time)
}
\arguments{
\item{vector}{A vector variable to be cleaned}

\item{time}{A vector variable indicating the elapsed time, needed to compute
velocity.}
}
\value{
A numeric vector processed as requested or as per the default. .
}
\description{
The function calls, in order, 'speed_clean', 'interpolate', and 'smooth_vector' from this packages.
Parameters can only be changed in the package options, i.e. 'Pupilla::pp_options()'.
Warning, the best preprocessing parameters may deviate from the defaults used here,
be mindful!
}
