% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TOBII.R
\name{read_TOBII}
\alias{read_TOBII}
\title{Reads and imports TOBII eye-tracking and behavioral data}
\usage{
read_TOBII(
  ID,
  path = getwd(),
  start_filename = "subject-",
  append_TOBII = "_TOBII_output.tsv",
  skip = 7,
  separate_behavioral = TRUE
)
}
\arguments{
\item{ID}{An integer corresponding to one participant's ID. This is attached
to the current path in order to locate the two files to read.
Also, One variable named p_ID is attached to both ET and BD files.
If a vector is supplied, files are read with \code{lapply} and then merged.}

\item{path}{Defaults to \code{getwd()} but can be specified to be otherwise.
Files will be searched for starting from this location.}

\item{start_filename}{A string, it defaults to \code{"subject-"}. Usually
all files start with this string, regardless of their nature. Usually
names are built by concatenating \code{path}, \code{start_filename},
\code{ID}, and \code{append_TOBII} (for eye-tracking data, else ".csv").}

\item{append_TOBII}{A string, it defaults to \code{"_TOBII_output.tsv"} and
indicates the text that tells eye-tracking from behavioral files apart. Usually
names are built by concatenating \code{path}, \code{start_filename},
\code{ID}, and \code{append_TOBII} (for eye-tracking data, else ".csv").}

\item{skip}{Integer. The amount of lines to skip from the eye-tracking file,
i.e. after how many lines the header is encountered.
This is passed to \code{data.table::fread()}.}

\item{separate_behavioral}{defaults to \code{TRUE}. If \code{FALSE}, it only reads
eye-tracking data}
}
\value{
A list of one or two DFs, one for eye-tracking data, one for behavioral data (if requested).
}
\description{
These two type of data are usually separate and sometimes must be merged
to create one single file with both eyetracking and experimental details
(e.g., conditions), unless you have specified this via code in the experiment builder.
This function is written with OpenSesame-like csv files in mind, though compatibility
with other programs (e.g., e-prime) may be achieved provided that files are converted to
the csv format. This is optimized for Windows machines - you may encounter
adress errors in Macs.
}
