% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_clean.R
\name{speed_clean}
\alias{speed_clean}
\title{Help identifying artifacts with a speed-based criterion}
\usage{
speed_clean(
  vector,
  time,
  thresh = pp_options("thresh"),
  speed_method = pp_options("speed_method"),
  extend_by = pp_options("extend_by"),
  island_size = pp_options("island_size")
)
}
\arguments{
\item{vector}{A vector variable to be cleaned}

\item{time}{A vector variable indicating the elapsed time, needed to compute
velocity.}

\item{thresh}{Threshold (z point or absolute value) above which values are marked as NA.}

\item{speed_method}{Whether the 'thresh' is a z-score ('z'), with deviant values
omitted once, or until there are no more values above the threshold ('z-dynamic').
'abs' is used instead when a precise absolute value for speed is supplied.}

\item{extend_by}{Number of samples starting from the deviant speed values that
are stripped (e.g., the signal in proximity of blinks may be biased as well).}

\item{island_size}{Islands of signal in the midst of NAs are removed if smaller
or equal to this threshold (amount of samples).}
}
\value{
A numeric vector cleaned as requested.
}
\description{
A signal vector is stripped of values exceeding a given
threshold, computed on the basis of the absolute speed of signal
increase or decrease.
}
