% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmem_test.R
\name{lmem_test}
\alias{lmem_test}
\title{Highlights and test the time-course of effects through mixed-effects models}
\usage{
lmem_test(data, formula, time, family = "gaussian", correction = "fdr", ...)
}
\arguments{
\item{data}{A data.frame containing all the necessary variables.}

\item{formula}{A 'lme4'-style formula, passed as a string.}

\item{time}{A string indicating the name of the time variable.}

\item{family}{A string indicating a GLM family to be passed to (g)lmer.}

\item{correction}{A string indicating the method for correcting p values (see p.adjust). defaults to "fdr".}

\item{...}{Other pars passed to (g)lmer.}
}
\value{
A list including: t/z values, p values, corrected p values.
}
\description{
This function performs (g)lmer for each timepoint.
Data are first separated for each time-point, and a LMEM as specified by the 'formula'
parameter, which is passed to 'lmerTest::lmer', is performed.
The function returns t/z values, p values, and corrected p values.
}
