% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_vector.R
\name{smooth_vector}
\alias{smooth_vector}
\title{Smooth a time series through cubic splines}
\usage{
smooth_vector(
  vector,
  time,
  spar = pp_options("spar"),
  disable_smoothing = pp_options("disable_smoothing")
)
}
\arguments{
\item{vector}{A vector variable to be smoothed.}

\item{time}{A vector variable indicating the elapsed time.}

\item{spar}{Smoothing factor as in 'smooth.spline()'.}

\item{disable_smoothing}{Whether you want to disable smoothing. This parameter
is FALSE by default; if you want to disable the smoothing, set it to TRUE.}
}
\value{
A numeric vector smoothed as requested.
}
\description{
This can be used as a simple smoothing function, in the case of eyetracking
data it can be used as a low-pass filter, useful to correct artifacts,
blinks, etc.
}
